requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
